# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/organizations/{organizationId}/categories": {
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Categories"
        ]
      },
      "get": {
        "operationId": "CategoriesController_findAll",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Categories"
        ]
      }
    },
    "/api/organizations/{organizationId}/categories/{categoryId}": {
      "get": {
        "operationId": "CategoriesController_findOne",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Categories"
        ]
      },
      "patch": {
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Categories"
        ]
      },
      "delete": {
        "operationId": "CategoriesController_remove",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Categories"
        ]
      }
    },
    "/api/organizations/{organizationId}/categories/reorder": {
      "patch": {
        "operationId": "CategoriesController_reorder",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderCategoriesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Categories"
        ]
      }
    },
...
    "/api/organizations/{organizationId}/credits": {
      "get": {
        "operationId": "CreditsController_getBalance",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Credits"
        ]
      }
    },
    "/api/organizations/{organizationId}/credits/packages": {
      "get": {
        "operationId": "CreditsController_getPackages",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Credits"
        ]
      }
    },
    "/api/organizations/{organizationId}/credits/packages/{slug}": {
      "get": {
        "operationId": "CreditsController_getPackage",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Credits"
        ]
      }
    },
    "/api/organizations/{organizationId}/credits/purchase": {
      "post": {
        "operationId": "CreditsController_purchaseCredits",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseCreditsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Credits"
        ]
      }
    },
    "/api/organizations/{organizationId}/credits/history": {
      "get": {
        "operationId": "CreditsController_getHistory",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter nach Zahlungsstatus",
            "schema": {
              "example": "completed",
              "type": "string",
              "enum": [
                "pending",
                "completed",
                "failed",
                "refunded"
              ]
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Startdatum für Filterung",
            "schema": {
              "example": "2024-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Enddatum für Filterung",
            "schema": {
              "example": "2024-12-31T23:59:59.000Z",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Seitennummer für Pagination",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Anzahl der Ergebnisse pro Seite",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Credits"
        ]
      }
    },
    "/api/organizations/{organizationId}/credits/licenses": {
      "get": {
        "operationId": "CreditsController_getLicenseUsage",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Startdatum für Filterung",
            "schema": {
              "example": "2024-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Enddatum für Filterung",
            "schema": {
              "example": "2024-12-31T23:59:59.000Z",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Seitennummer für Pagination",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Anzahl der Ergebnisse pro Seite",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Credits"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenEOS API",
    "description": "\n# OpenEOS API Documentation\n\nOpen Event Ordering System - REST API für Kassensysteme bei Vereinen und Veranstaltungen.\n\n## Authentifizierung\n\nDie API verwendet JWT Bearer Tokens für die Authentifizierung. Nach dem Login erhältst du einen Access Token (30 Min gültig) und einen Refresh Token (7 Tage, als httpOnly Cookie).\n\n### Endpoints ohne Authentifizierung\n- `POST /api/auth/register` - Registrierung\n- `POST /api/auth/login` - Login\n- `POST /api/auth/refresh` - Token erneuern\n- `POST /api/auth/forgot-password` - Passwort vergessen\n- `POST /api/auth/reset-password` - Passwort zurücksetzen\n- `GET /api/health/*` - Health Checks\n- `GET /api/invitations/:token` - Einladung abrufen\n- `/api/public/order/*` - Online-Bestellungen (mit Session-Token)\n\n## Multi-Tenant\n\nDie API ist multi-tenant-fähig. Die meisten Endpoints erfordern eine `organizationId` im Pfad.\n\n## Rollen\n\n- **ADMIN** - Vollzugriff auf Organisation\n- **MANAGER** - Verwaltung von Produkten, Events, Bestellungen\n- **CASHIER** - Kassen-Operationen\n- **KITCHEN** - Küchen-Display\n- **DELIVERY** - Ausgabe-Display\n\n## WebSocket\n\nReal-time Updates über Socket.io unter `/` mit Namespaces:\n- `/pos` - Kassen-Clients\n- `/display` - Küchen/Ausgabe-Displays\n- `/printer` - Drucker-Agenten\n- `/admin` - Admin-Dashboard\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "OpenEOS",
...
    },
    "license": {
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "X-Organization-Id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Organization-Id",
        "description": "Organization ID for multi-tenant requests"
      },
      "X-Session-Token": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Session-Token",
        "description": "Session token for online orders"
      }
    },
    "schemas": {
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Hauptgerichte",
            "description": "Name der Kategorie"
          },
          "description": {
            "type": "string",
            "example": "Warme Hauptspeisen und Gerichte",
            "description": "Beschreibung der Kategorie"
          },
          "parentId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "description": "ID der übergeordneten Kategorie"
          },
          "color": {
            "type": "string",
            "example": "#FF5733",
            "description": "Farbcode der Kategorie (Hex-Format)"
          },
          "icon": {
            "type": "string",
            "example": "restaurant",
            "description": "Icon-Name für die Kategorie"
          },
          "sortOrder": {
            "type": "number",
            "example": 1,
            "description": "Sortierreihenfolge"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Gibt an, ob die Kategorie aktiv ist"
          },
          "printSettings": {
            "type": "object",
            "example": {
              "printToKitchen": true,
              "printerIds": []
            },
            "description": "Druckeinstellungen für die Kategorie"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Hauptgerichte",
            "description": "Name der Kategorie"
          },
          "description": {
            "type": "string",
            "example": "Warme Hauptspeisen und Gerichte",
            "description": "Beschreibung der Kategorie"
          },
          "parentId": {
            "type": "object",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "description": "ID der übergeordneten Kategorie"
          },
          "color": {
            "type": "object",
            "example": "#FF5733",
            "description": "Farbcode der Kategorie (Hex-Format)"
          },
          "icon": {
            "type": "object",
            "example": "restaurant",
            "description": "Icon-Name für die Kategorie"
          },
          "sortOrder": {
            "type": "number",
            "example": 1,
            "description": "Sortierreihenfolge"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Gibt an, ob die Kategorie aktiv ist"
          },
          "printSettings": {
            "type": "object",
            "example": {
              "printToKitchen": true,
              "printerIds": []
            },
            "description": "Druckeinstellungen für die Kategorie"
          }
        }
      },
      "ReorderCategoriesDto": {
        "type": "object",
        "properties": {
          "categoryIds": {
            "example": [
              "550e8400-e29b-41d4-a716-446655440000",
              "550e8400-e29b-41d4-a716-446655440001"
            ],
            "description": "Sortierte Liste der Kategorie-IDs",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PurchaseCreditsDto": {
        "type": "object",
        "properties": {
          "packageId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "description": "ID des Credit-Pakets"
          },
          "paymentMethod": {
            "type": "string",
            "example": "stripe",
            "description": "Zahlungsmethode",
            "enum": [
              "sumup_online",
              "bank_transfer",
              "invoice"
            ]
          }
        },
        "required": [
          "packageId",
          "paymentMethod"
        ]
      },
...
    }
  }
}